#381: remove class name from test stages
